+++ tests/e2e-frontend/.gitignore
@@ -1,3 +1,4 @@
 /node_modules/
 /playwright-report/
 /test-results/
+users.json
