+++ tests/e2e-frontend/tests/users.template.json
@@ -1,0 +1,12 @@
+{
+    "osparc": [{
+        "email": "email.osparc@domain.com",
+        "password": "my-password",
+        "role": "user"
+    }],
+    "s4l": [{
+        "email": "email.s4l@domain.com",
+        "password": "my-password",
+        "role": "user"
+    }]
+}
